# crypto: optimize SipHash Write() method with chunked processing