@l0rinc 2025-09-06
This seems like a good idea to me, I will review this in more detail soon.

To set your expectations, these fundamental changes can take a long time to review - and we need very good reasons to risk introducing a potential error and making the code more complicated for a speedup that the users might never notice.

I like that you're explaining it in a lot of detail in the commit message, it's not a trivial change! We may be able to use more C++20 alignment-specific methods there (e.g. like the `std::assume_aligned` we used in https://github.com/bitcoin/bitcoin/pull/31144/files#diff-f26d4597a7f5a5d4aa30053032d49427b402ef4fd7a1b3194cb75b2551d58ca4R44) - but I haven't gone deep yet, I could be wrong here, I just glanced at it so far.

In the meantime, please check if there's any way to split this into even smaller commits to make it even easier for reviewers, check if the current tests cover each branch (e.g. the big-endian ones), different architectures, etc.

But as far as I can tell this doesn't affect most caching scenarios since we have dedicated siphash methods for those.
To be sure, I can run some reindexes once the change is settled to make sure it doesn't make it slower at least.