@Raimo33 2025-09-07
I just realized somebody had my same idea on knots28. But in the latest version it got reset to match Core's implementation. 🤔.
https://github.com/Raimo33/bitcoinknots/commit/8463aef09fe4c6e3806e3b8a4c967c297302b96f