@Raimo33 2025-09-07
> We may be able to use more C++20 alignment-specific methods there (e.g. like the `std::assume_aligned` we used in

I tried with
```diff
- uint64_t chunk;
- std::memcpy(&chunk, ptr, 8);
+ auto aligned_ptr = std::assume_aligned<8>(ptr);
+ uint64_t chunk = *reinterpret_cast<const uint64_t *>(aligned_ptr);
```
but no luck. performance is identical. `memcpy` is extremely optimized.