@Raimo33 2025-09-17
> In the meantime, please check if there's any way to split this into even smaller commits to make it even easier for reviewers

I've just split the PR into 3 commits, diff should now be simpler