@luke-jr 2025-09-18
> I just realized this was attempted in the past. merged in Bitcoin knots 19. restored to match Core's implementation in knots 29. 🤔.

#18014 is still in Knots 29...